chore: securityConfig 파일 중복 제거
